re-implemented v1 and v2 of workflow using args
